Fix routing topk_indices dtype (int64→int32) and add test progress logs
- moe_routing.py: topk_indices must be int32 for fused_topk_deepseek C++ kernel
- test_cute_dsl_grouped_gemm_fp8.py: add per-case progress prints
- test_cute_dsl_agather_gemm.py: add per-case progress prints

+++ flashinfer/cute_dsl/moe_routing.py
@@ -145,7 +145,7 @@ def moe_routing_deepseek(
 
     # Allocate output tensors for fused routing
     topk_values = torch.empty(T, top_k, dtype=torch.float32, device=device)
-    topk_indices = torch.empty(T, top_k, dtype=torch.int64, device=device)
+    topk_indices = torch.empty(T, top_k, dtype=torch.int32, device=device)
 
     # Run fused DeepSeek-V3 routing kernel
     fused_topk_deepseek(
